requirements.txt
Flask
numpy
tensorflow
cs50
Flask-Session
pytz
requests
os
torch
keras2onnx
tf2onnx
onnxruntime
onnx
pillow
io
base64
datetime
Werkzeug